requirements.txt
alabaster==0.7.13
asgiref==3.7.2
Babel==2.12.1
beautifulsoup4==4.12.2
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
crispy-bootstrap5==2023.10
defusedxml==0.7.1
Django==4.2.4
django-crispy-forms==2.0
docutils==0.20.1
gitdb==4.0.10
GitPython==3.1.34
gunicorn==21.2.0
idna==3.4
imagesize==1.4.1
isort==5.12.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
pguard==0.4.0
Pillow==10.0.1
platformdirs==3.11.0
psycopg==3.1.10
Pygments==2.16.1
requests==2.31.0
setuptools-scm==7.1.0
smmap==5.0.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.5
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.9.0
urllib3==2.0.4
